Rename package info parser
