+++ cppwg/generators.py
@@ -15,7 +15,7 @@ from cppwg.input.class_info import CppClassInfo
 from cppwg.input.free_function_info import CppFreeFunctionInfo
 from cppwg.input.info_helper import CppInfoHelper
 from cppwg.input.package_info import PackageInfo
-from cppwg.parsers.package_info import PackageInfoParser
+from cppwg.parsers.package_info_parser import PackageInfoParser
 from cppwg.parsers.source_parser import CppSourceParser
 from cppwg.templates import pybind11_default as wrapper_templates
 from cppwg.utils.constants import CPPWG_EXT, CPPWG_HEADER_COLLECTION_FILENAME
