+++ cppwg/parsers/package_info_parser.py
